refactor: Rename directory name 'domain' to 'entity'
