+++ src/main/java/com/smplatform/member_service/domain/member/repository/CustomMemberRepository.java
@@ -1,6 +1,6 @@
 package com.smplatform.member_service.domain.member.repository;
 
-import com.smplatform.member_service.domain.member.domain.Member;
+import com.smplatform.member_service.domain.member.entity.Member;
 import com.smplatform.member_service.domain.member.dto.MemberSearchRequestParamDto;
 
 import java.util.List;
