+++ src/main/java/com/smplatform/member_service/domain/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.smplatform.member_service.domain.member.repository;
 
-import com.smplatform.member_service.domain.member.domain.Member;
+import com.smplatform.member_service.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
